Move settings-related components to "settings" folder inside "components"
