+++ frontend/src/components/settings/SettingsSwitch.tsx
