+++ frontend/src/pages/protected/Play.tsx
@@ -14,7 +14,7 @@ import {
 } from "../../utils/apiUtils.ts";
 
 import GameOverModal from "../../components/global/modals/gameOverModals/MultiplayerModal.tsx";
-import GameplaySettings from "../../features/modals/GameplaySettings/GameplaySettings.tsx";
+import GameplaySettings from "../../features/modals/GameplaySettings.tsx";
 import ModalWrapper from "../../components/global/wrappers/ModalWrapper.js";
 import { OptionalValue } from "../../types/general.js";
 import {
